5 aula

+++ 04-JPA-DAO/src/main/java/br/com/fiap/dao/impl/ImovelDAOImpl.java
@@ -5,6 +5,7 @@ import javax.persistence.EntityManager;
 import br.com.fiap.dao.ImovelDAO;
 import br.com.fiap.entity.Imovel;
 import br.com.fiap.excecao.CommitException;
+import br.com.fiap.excecao.SearchNotFoundException;
 
 public class ImovelDAOImpl implements ImovelDAO{
 	
@@ -27,8 +28,10 @@ public class ImovelDAOImpl implements ImovelDAO{
 		em.merge(imovel);
 	}
 
-	public void remover(int codigo) {
+	public void remover(int codigo) throws SearchNotFoundException {
 		Imovel imovel = consultar(codigo);
+		if(imovel==null)
+			throw new SearchNotFoundException();
 		em.remove(imovel);		
 	}
